Move out www extra link tests
