+++ tests/www/test_views.py
@@ -44,7 +44,6 @@ from airflow.configuration import conf
 from airflow.executors.celery_executor import CeleryExecutor
 from airflow.jobs.base_job import BaseJob
 from airflow.models import DAG, DagRun, TaskInstance
-from airflow.models.baseoperator import BaseOperator, BaseOperatorLink
 from airflow.models.renderedtifields import RenderedTaskInstanceFields as RTIF
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.operators.bash import BashOperator
@@ -2847,219 +2846,3 @@ class TestTriggerDag(TestBase):
         resp = self.client.get(url, follow_redirects=True)
         response_data = resp.data.decode()
         assert "Access is Denied" in response_data
-
-
-class TestExtraLinks(TestBase):
-    def setUp(self):
-        from tests.test_utils.mock_operators import Dummy2TestOperator, Dummy3TestOperator
-
-        self.endpoint = "extra_links"
-        self.default_date = datetime(2017, 1, 1)
-
-        class RaiseErrorLink(BaseOperatorLink):
-            name = 'raise_error'
-
-            def get_link(self, operator, dttm):
-                raise ValueError('This is an error')
-
-        class NoResponseLink(BaseOperatorLink):
-            name = 'no_response'
-
-            def get_link(self, operator, dttm):  # pylint: disable=unused-argument
-                return None
-
-        class FooBarLink(BaseOperatorLink):
-            name = 'foo-bar'
-
-            def get_link(self, operator, dttm):
-                return f"http://www.example.com/{operator.task_id}/foo-bar/{dttm}"
-
-        class AirflowLink(BaseOperatorLink):
-            name = 'airflow'
-
-            def get_link(self, operator, dttm):  # pylint: disable=unused-argument
-                return 'https://airflow.apache.org'
-
-        class DummyTestOperator(BaseOperator):
-
-            operator_extra_links = (
-                RaiseErrorLink(),
-                NoResponseLink(),
-                FooBarLink(),
-                AirflowLink(),
-            )
-
-        self.dag = DAG('dag', start_date=self.default_date)
-        self.task = DummyTestOperator(task_id="some_dummy_task", dag=self.dag)
-        self.task_2 = Dummy2TestOperator(task_id="some_dummy_task_2", dag=self.dag)
-        self.task_3 = Dummy3TestOperator(task_id="some_dummy_task_3", dag=self.dag)
-
-        self.app.dag_bag = mock.MagicMock(**{'get_dag.return_value': self.dag})
-        super().setUp()
-        self.logout()
-        self.login('test_viewer', 'test_viewer')
-
-    def test_extra_links_works(self):
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=foo-bar".format(
-                self.endpoint, self.dag.dag_id, self.task.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {
-            'url': 'http://www.example.com/some_dummy_task/foo-bar/2017-01-01T00:00:00+00:00',
-            'error': None,
-        }
-
-    def test_global_extra_links_works(self):
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=github".format(
-                self.endpoint, self.dag.dag_id, self.task.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://github.com/apache/airflow', 'error': None}
-
-    def test_extra_link_in_gantt_view(self):
-        exec_date = dates.days_ago(2)
-        start_date = datetime(2020, 4, 10, 2, 0, 0)
-        end_date = exec_date + timedelta(seconds=30)
-
-        with create_session() as session:
-            for task in self.dag.tasks:
-                ti = TaskInstance(task=task, execution_date=exec_date, state="success")
-                ti.start_date = start_date
-                ti.end_date = end_date
-                session.add(ti)
-
-        url = f'gantt?dag_id={self.dag.dag_id}&execution_date={exec_date}'
-        resp = self.client.get(url, follow_redirects=True)
-
-        self.check_content_in_response('"extraLinks":', resp)
-
-        extra_links_grps = re.search(r'extraLinks\": \[(\".*?\")\]', resp.get_data(as_text=True))
-        extra_links = extra_links_grps.group(0)
-        assert 'airflow' in extra_links
-        assert 'github' in extra_links
-
-    def test_operator_extra_link_override_global_extra_link(self):
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=airflow".format(
-                self.endpoint, self.dag.dag_id, self.task.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://airflow.apache.org', 'error': None}
-
-    def test_extra_links_error_raised(self):
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=raise_error".format(
-                self.endpoint, self.dag.dag_id, self.task.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert 404 == response.status_code
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': None, 'error': 'This is an error'}
-
-    def test_extra_links_no_response(self):
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=no_response".format(
-                self.endpoint, self.dag.dag_id, self.task.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 404
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': None, 'error': 'No URL found for no_response'}
-
-    def test_operator_extra_link_override_plugin(self):
-        """
-        This tests checks if Operator Link (AirflowLink) defined in the Dummy2TestOperator
-        is overridden by Airflow Plugin (AirflowLink2).
-
-        AirflowLink returns 'https://airflow.apache.org/' link
-        AirflowLink2 returns 'https://airflow.apache.org/1.10.5/' link
-        """
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=airflow".format(
-                self.endpoint, self.dag.dag_id, self.task_2.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://airflow.apache.org/1.10.5/', 'error': None}
-
-    def test_operator_extra_link_multiple_operators(self):
-        """
-        This tests checks if Operator Link (AirflowLink2) defined in
-        Airflow Plugin (AirflowLink2) is attached to all the list of
-        operators defined in the AirflowLink2().operators property
-
-        AirflowLink2 returns 'https://airflow.apache.org/1.10.5/' link
-        GoogleLink returns 'https://www.google.com'
-        """
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=airflow".format(
-                self.endpoint, self.dag.dag_id, self.task_2.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://airflow.apache.org/1.10.5/', 'error': None}
-
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=airflow".format(
-                self.endpoint, self.dag.dag_id, self.task_3.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://airflow.apache.org/1.10.5/', 'error': None}
-
-        # Also check that the other Operator Link defined for this operator exists
-        response = self.client.get(
-            "{}?dag_id={}&task_id={}&execution_date={}&link_name=google".format(
-                self.endpoint, self.dag.dag_id, self.task_3.task_id, self.default_date
-            ),
-            follow_redirects=True,
-        )
-
-        assert response.status_code == 200
-        response_str = response.data
-        if isinstance(response.data, bytes):
-            response_str = response_str.decode()
-        assert json.loads(response_str) == {'url': 'https://www.google.com', 'error': None}
